refactor: keep mark all as read mutation in component
