+++ src/app/Scenes/Activity/ActivityList.tsx
@@ -1,7 +1,5 @@
-import { captureException } from "@sentry/react-native"
 import { ActivityList_me$key } from "__generated__/ActivityList_me.graphql"
 import { ActivityList_viewer$key } from "__generated__/ActivityList_viewer.graphql"
-import { ActivityListMarkAllAsReadMutation } from "__generated__/ActivityListMarkAllAsReadMutation.graphql"
 import { ActivityQuery } from "__generated__/ActivityQuery.graphql"
 import { StickTabPageRefreshControl } from "app/Components/StickyTabPage/StickTabPageRefreshControl"
 import {
@@ -11,15 +9,8 @@ import {
 } from "app/Components/StickyTabPage/StickyTabPageFlatList"
 import { extractNodes } from "app/utils/extractNodes"
 import { Flex, Separator, Spinner } from "palette"
-import { useCallback, useContext, useEffect, useState } from "react"
-import {
-  ConnectionHandler,
-  graphql,
-  useFragment,
-  useMutation,
-  usePaginationFragment,
-} from "react-relay"
-import { RecordSourceSelectorProxy } from "relay-runtime"
+import { useContext, useEffect, useState } from "react"
+import { graphql, useFragment, usePaginationFragment } from "react-relay"
 import { ActivityEmptyView } from "./ActivityEmptyView"
 import { ActivityItem } from "./ActivityItem"
 import { ActivityMarkAllAsReadSection } from "./ActivityMarkAllAsReadSection"
@@ -35,8 +26,6 @@ interface ActivityListProps {
 export const ActivityList: React.FC<ActivityListProps> = ({ viewer, type, me }) => {
   const [refreshing, setRefreshing] = useState(false)
   const setJSX = useContext(StickyTabPageFlatListContext).setJSX
-  const [commit, mutationInProgress] =
-    useMutation<ActivityListMarkAllAsReadMutation>(MarkAllAsReadMutation)
   const { data, hasNext, isLoadingNext, loadNext, refetch } = usePaginationFragment<
     ActivityQuery,
     ActivityList_viewer$key
@@ -79,42 +68,14 @@ export const ActivityList: React.FC<ActivityListProps> = ({ viewer, type, me }) 
     )
   }
 
-  const handleMarkAllAsReadPress = useCallback(() => {
-    try {
-      commit({
-        variables: {},
-        updater: markAllAsReadMutationUpdater,
-        optimisticUpdater: markAllAsReadMutationUpdater,
-        onCompleted: (response) => {
-          const errorMessage =
-            response.markAllNotificationsAsRead?.responseOrError?.mutationError?.message
-
-          if (errorMessage) {
-            throw new Error(errorMessage)
-          }
-        },
-      })
-    } catch (e) {
-      if (__DEV__) {
-        console.error(e)
-      } else {
-        captureException(e)
-      }
-    }
-  }, [])
-
   useEffect(() => {
     setJSX(
       <>
-        <ActivityMarkAllAsReadSection
-          hasUnreadNotifications={hasUnreadNotifications}
-          loading={mutationInProgress}
-          onPress={handleMarkAllAsReadPress}
-        />
+        <ActivityMarkAllAsReadSection hasUnreadNotifications={hasUnreadNotifications} />
         <Separator />
       </>
     )
-  }, [hasUnreadNotifications, mutationInProgress, handleMarkAllAsReadPress])
+  }, [hasUnreadNotifications])
 
   if (notifications.length === 0) {
     return <ActivityEmptyView type={type} />
@@ -169,43 +130,3 @@ const meFragment = graphql`
     unreadNotificationsCount
   }
 `
-
-const MarkAllAsReadMutation = graphql`
-  mutation ActivityListMarkAllAsReadMutation {
-    markAllNotificationsAsRead(input: {}) {
-      responseOrError {
-        ... on MarkAllNotificationsAsReadSuccess {
-          success
-        }
-        ... on MarkAllNotificationsAsReadFailure {
-          mutationError {
-            message
-          }
-        }
-      }
-    }
-  }
-`
-
-const markAllAsReadMutationUpdater = (store: RecordSourceSelectorProxy) => {
-  const root = store.getRoot()
-  const me = root.getLinkedRecord("me")
-  const viewer = root.getLinkedRecord("viewer")
-
-  if (!me || !viewer) {
-    return
-  }
-
-  const key = "ActivityList_notificationsConnection"
-  const connection = ConnectionHandler.getConnection(viewer, key)
-  const edges = connection?.getLinkedRecords("edges")
-
-  // Set unread notifications count to 0
-  me.setValue(0, "unreadNotificationsCount")
-
-  // Mark all notifications as read
-  edges?.forEach((edge) => {
-    const node = edge.getLinkedRecord("node")
-    node?.setValue(false, "isUnread")
-  })
-}
